* commit '95b5999a68b018e0bcf0cb4fe89f704a578254b3':
  add file modified time

+++ client/src/main/java/com/alibaba/nacos/client/config/impl/LocalConfigInfoProcessor.java
@@ -97,6 +97,11 @@ public class LocalConfigInfoProcessor {
         }
         File file = getSnapshotFile(envName, dataId, group, tenant);
         if (null == config) {
+            if (JVMUtil.isMultiInstance() && file.exists() && file.lastModified() >= ConcurrentDiskUtil.FILELASTMODIFIED.get()) {
+                ConcurrentDiskUtil.FILELASTMODIFIED.remove();
+                return ;
+
+            }
             try {
                 IOUtils.delete(file);
             } catch (IOException ioe) {
